+++ partials/scroll-top.php
@@ -16,7 +16,7 @@ if ( true != get_theme_mod( 'ocean_scroll_top', true ) ) {
 }
 
 // Get arrow
-$arrow = get_theme_mod( 'ocean_scroll_top_arrow' );
+$arrow = apply_filters( 'ocean_scroll_top_arrow', get_theme_mod( 'ocean_scroll_top_arrow' ) );
 $arrow = $arrow ? $arrow : 'fa fa-angle-up'; ?>
 
 <a id="scroll-top" href="#"><span class="<?php echo esc_attr( $arrow ); ?>"></span></a>