+++ sidebar.php
@@ -3,7 +3,12 @@
  * The sidebar containing the main widget area.
  *
  * @package OceanWP WordPress theme
- */ ?>
+ */
+
+ // Retunr if full width or full screen
+if ( in_array( oceanwp_post_layout(), array( 'full-screen', 'full-width' ) ) ) {
+	return;
+} ?>
 
 <?php do_action( 'ocean_before_sidebar' ); ?>